# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Email Microservice API",
    "description": "API untuk Email Microservice yang mendukung pengiriman email, manajemen template, dan pelacakan email.",
    "version": "1.0.0",
    "contact": {
      "name": "Tim API Email Microservice",
      "email": "api-support@example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Development Server"
    },
    {
      "url": "https://api.example.com/api/v1",
      "description": "Production Server"
    }
  ],
...
  "paths": {
...
    "/emails/template": {
      "post": {
        "tags": ["Email"],
        "summary": "Mengirim email dengan template",
        "description": "Mengirim email menggunakan template yang sudah ada dengan variabel yang ditentukan",
        "operationId": "sendTemplateEmail",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "description": "Payload pengiriman email dengan template",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TemplateEmailRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Email diantrekan untuk pengiriman",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validasi gagal",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Autentikasi gagal",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Template tidak ditemukan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit terlampaui",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/templates": {
      "post": {
        "tags": ["Template"],
        "summary": "Membuat template email baru",
        "description": "Membuat template email baru untuk digunakan dalam pengiriman email",
        "operationId": "createTemplate",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "description": "Payload pembuatan template",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TemplateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Template berhasil dibuat",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validasi gagal",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Autentikasi gagal",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Template dengan nama tersebut sudah ada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Template"],
        "summary": "Mendapatkan daftar template",
        "description": "Mendapatkan daftar template email yang tersedia",
        "operationId": "getTemplates",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Nomor halaman",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Jumlah item per halaman",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "category",
            "in": "query",
            "description": "Filter berdasarkan kategori",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filter berdasarkan status",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["active", "inactive"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Daftar template berhasil diambil",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplatesListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Autentikasi gagal",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/templates/{id}": {
      "get": {
        "tags": ["Template"],
        "summary": "Mendapatkan detail template",
        "description": "Mendapatkan detail lengkap template email",
        "operationId": "getTemplate",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID template",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detail template berhasil diambil",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateDetailResponse"
                }
              }
            }
          },
          "401": {
            "description": "Autentikasi gagal",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Template tidak ditemukan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Template"],
        "summary": "Memperbarui template",
        "description": "Memperbarui template email yang sudah ada",
        "operationId": "updateTemplate",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID template",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Payload pembaruan template",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TemplateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Template berhasil diperbarui",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateUpdateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validasi gagal",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Autentikasi gagal",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Template tidak ditemukan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Template"],
        "summary": "Menghapus template",
        "description": "Menghapus template email",
        "operationId": "deleteTemplate",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID template",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Template berhasil dihapus",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "401": {
            "description": "Autentikasi gagal",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Template tidak ditemukan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TemplateEmailRequest": {
        "type": "object",
        "required": ["templateId", "to", "variables"],
        "properties": {
          "templateId": {
            "type": "string",
            "description": "ID template yang akan digunakan"
          },
          "to": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "email"
            },
            "description": "Daftar penerima email"
          },
          "cc": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "email"
            },
            "description": "Daftar carbon copy"
          },
          "bcc": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "email"
            },
            "description": "Daftar blind carbon copy"
          },
          "variables": {
            "type": "object",
            "additionalProperties": true,
            "description": "Variabel yang akan digunakan dalam template"
          },
          "attachments": {
            "type": "array",
            "items": {
              "type": "object",
              "required": ["filename", "content", "contentType"],
              "properties": {
                "filename": {
                  "type": "string",
                  "description": "Nama file lampiran"
                },
                "content": {
                  "type": "string",
                  "description": "Konten file dalam format base64"
                },
                "contentType": {
                  "type": "string",
                  "description": "MIME type file"
                }
              }
            }
          },
          "metadata": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": "Metadata untuk keperluan tracking"
          }
        }
      },
      "EmailResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["success"]
          },
          "traceID": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "properties": {
              "email_id": {
                "type": "string",
                "description": "ID unik email"
              }
            }
          }
        }
      },
...
      "TemplateRequest": {
        "type": "object",
        "required": ["name", "subject", "htmlContent"],
        "properties": {
          "name": {
            "type": "string",
            "description": "Nama unik untuk template"
          },
          "subject": {
            "type": "string",
            "description": "Subjek email dengan placeholder"
          },
          "description": {
            "type": "string",
            "description": "Deskripsi template"
          },
          "htmlContent": {
            "type": "string",
            "description": "Konten HTML dengan placeholder"
          },
          "textContent": {
            "type": "string",
            "description": "Versi teks biasa"
          },
          "category": {
            "type": "string",
            "description": "Kategori untuk pengelompokan"
          },
          "isActive": {
            "type": "boolean",
            "default": true,
            "description": "Status template"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true,
            "description": "Metadata tambahan"
          }
        }
      },
      "TemplateResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["success"]
          },
          "traceID": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "description": "ID template"
              },
              "name": {
                "type": "string",
                "description": "Nama template"
              },
              "subject": {
                "type": "string",
                "description": "Subjek email"
              },
              "createdAt": {
                "type": "string",
                "format": "date-time",
                "description": "Waktu pembuatan"
              },
              "updatedAt": {
                "type": "string",
                "format": "date-time",
                "description": "Waktu pembaruan terakhir"
              }
            }
          }
        }
      },
      "TemplatesListResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["success"]
          },
          "traceID": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "properties": {
              "templates": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "description": "ID template"
                    },
                    "name": {
                      "type": "string",
                      "description": "Nama template"
                    },
                    "subject": {
                      "type": "string",
                      "description": "Subjek email"
                    },
                    "description": {
                      "type": "string",
                      "description": "Deskripsi template"
                    },
                    "category": {
                      "type": "string",
                      "description": "Kategori template"
                    },
                    "isActive": {
                      "type": "boolean",
                      "description": "Status template"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "description": "Waktu pembuatan"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time",
                      "description": "Waktu pembaruan terakhir"
                    }
                  }
                }
              },
              "pagination": {
                "type": "object",
                "properties": {
                  "currentPage": {
                    "type": "integer",
                    "description": "Halaman saat ini"
                  },
                  "totalPages": {
                    "type": "integer",
                    "description": "Total halaman"
                  },
                  "totalItems": {
                    "type": "integer",
                    "description": "Total item"
                  },
                  "itemsPerPage": {
                    "type": "integer",
                    "description": "Jumlah item per halaman"
                  }
                }
              }
            }
          }
        }
      },
      "TemplateDetailResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["success"]
          },
          "traceID": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "description": "ID template"
              },
              "name": {
                "type": "string",
                "description": "Nama template"
              },
              "subject": {
                "type": "string",
                "description": "Subjek email"
              },
              "description": {
                "type": "string",
                "description": "Deskripsi template"
              },
              "htmlContent": {
                "type": "string",
                "description": "Konten HTML"
              },
              "textContent": {
                "type": "string",
                "description": "Konten teks biasa"
              },
              "category": {
                "type": "string",
                "description": "Kategori template"
              },
              "isActive": {
                "type": "boolean",
                "description": "Status template"
              },
              "metadata": {
                "type": "object",
                "additionalProperties": true,
                "description": "Metadata template"
              },
              "createdAt": {
                "type": "string",
                "format": "date-time",
                "description": "Waktu pembuatan"
              },
              "updatedAt": {
                "type": "string",
                "format": "date-time",
                "description": "Waktu pembaruan terakhir"
              }
            }
          }
        }
      },
      "TemplateUpdateResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["success"]
          },
          "traceID": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "description": "ID template"
              },
              "updatedAt": {
                "type": "string",
                "format": "date-time",
                "description": "Waktu pembaruan"
              }
            }
          }
        }
      },
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["success"]
          },
          "traceID": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "null"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["error"]
          },
          "traceID": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "properties": {
              "errors": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "field": {
                      "type": "string",
                      "description": "Nama field yang error"
                    },
                    "message": {
                      "type": "string",
                      "description": "Deskripsi error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "AdminTokenAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Admin-Token"
      }
    }
  }
}